+++ libexec/scoop-depends.ps1
@@ -2,6 +2,8 @@
 # Summary: List dependencies for application(s).
 # Help: All dependencies will be "resolved"/checked, if they are accessible (in case of remote manifests or different versions).
 #
+# Supports the same format of <APP> parameter as in "scoop install" (See: 'scoop install --help')
+#
 # If application was already resolved as a dependency, duplicate will not be added (even when the versions are different).
 # 'shovel depends 7zip lessmsi@1.9.0' will resolve just to `main/lessmsi` instead of 'main/lessmsi main/lessmsi@1.9.0'
 # Output of depends command could be used for 'shovel cat' command to verify and check all the manifests, which will be installed.